Update structure rules in config and test files
The commit modifies the structure rules in `test_config.yaml` file making it more specific for Python packages. A new pytest was added to `repo_structure_test.py` to ensure failure of parsing structure rules when there includes are dangling. In addition, the `python_package` rule in the tests was also updated to reflect the changes made in the config file.